personaplex/requirements.txt
fastapi
uvicorn[standard]
telnyx
python-dotenv
python-multipart
httpx
websockets
sphn
numpy
noisereduce
